Update HAL Mock modules (#31) and add unit tests for Nonin OEM III (#80) (#177)
* Updated HAL Mock modules (#31) and added TestCache2 test scenarios for Nonin OEM III(#80)

* Updated file Headers for mock modules and their interfaces.
Updated file headers for HalAnalogInput and HALSPIDevice.

* Resolved file header format update for AnalogInput.h

* Resolved static check for TestCache2 files.

* Resolved Mock Modules static checks errors.

* Fixed issue with TestCache2 build.

* updated .cproject file for Testcache2 compilation.

* Fixed Issue with Interfaces file inclusion.

* Resolved Static checks for mock modules and Testcache2 files.

* Resolved static checks for Nonin OEM III.

* Updated the test frame receiver for checksum validation.

* Apply clang-format autofixes

* Resolve problems with building Catch2 tests

* Apply clang-tidy autofixes

* Resolve remaining clang-tidy issues

* Add LLVM repository key

+++ firmware/ventilator-controller-stm32/Core/Src/Pufferfish/Driver/Serial/Nonin/FrameBuffer.cpp
@@ -66,7 +66,7 @@ void FrameBuffer::shift_left() {
   /* On no frame data available frameBuffer and frameIndex are not updated */
   if (received_length_ > 0) {
     /* Update the frame buffer and index to receive new byte data */
-    for (size_t index = 0; index < (received_length_ - 1); index++) {
+    for (size_t index = 0; index < static_cast<size_t>(received_length_ - 1); index++) {
       frame_buffer_[index] = frame_buffer_[index + 1];
     }
     received_length_--;
